rename: UriProvider -> UriGenerator 로 이름 변경
