+++ src/main/java/com/commerce/backendserver/newauth/domain/OAuthUriGenerator.java
@@ -2,7 +2,7 @@ package com.commerce.backendserver.newauth.domain;
 
 import com.commerce.backendserver.newauth.domain.model.OAuthProvider;
 
-public interface OAuthUriProvider {
+public interface OAuthUriGenerator {
 
 	boolean isSupported(final OAuthProvider provider);
